Move config to common
