+++ bin/strata-client/src/l1_reader.rs
@@ -1,6 +1,10 @@
 use std::sync::Arc;
 
-use strata_btcio::{reader::query::bitcoin_data_reader_task, rpc::traits::Reader};
+use strata_btcio::{
+    reader::{config::ReaderConfig, query::bitcoin_data_reader_task},
+    rpc::traits::Reader,
+};
+use strata_common::config::Config;
 use strata_consensus_logic::{csm::ctl::CsmController, l1_handler::bitcoin_data_handler_task};
 use strata_db::traits::{Database, L1Database};
 use strata_primitives::params::Params;
@@ -8,8 +12,6 @@ use strata_status::StatusChannel;
 use strata_tasks::TaskExecutor;
 use strata_tx_parser::messages::L1Event;
 use tokio::sync::mpsc;
-
-use crate::config::Config;
 
 pub fn start_reader_tasks<D>(
     executor: &TaskExecutor,
@@ -31,7 +33,10 @@ where
     let target_next_block = l1_db.get_chain_tip()?.map(|i| i + 1).unwrap_or(horz_height);
     assert!(target_next_block >= horz_height);
 
-    let reader_config = Arc::new(config.get_reader_config(params.clone()));
+    let reader_config = Arc::new(ReaderConfig::from_config_and_params(
+        config.clone(),
+        params.clone(),
+    ));
 
     executor.spawn_critical_async(
         "bitcoin_data_reader_task",
