+++ bin/strata-client/src/main.rs
@@ -1,7 +1,6 @@
 use std::{str::FromStr, sync::Arc, time::Duration};
 
 use bitcoin::{hashes::Hash, Address, BlockHash};
-use config::{ClientMode, Config, SequencerConfig};
 use jsonrpsee::Methods;
 use rpc_client::sync_client;
 use strata_bridge_relay::relayer::RelayerHandle;
@@ -10,7 +9,10 @@ use strata_btcio::{
     rpc::{traits::Reader, BitcoinClient},
     writer::{config::WriterConfig, start_inscription_task},
 };
-use strata_common::logging;
+use strata_common::{
+    config::{ClientMode, Config, SequencerConfig},
+    logging,
+};
 use strata_consensus_logic::{
     checkpoint::CheckpointHandle,
     duty::{types::DutyBatch, worker as duty_worker},
@@ -27,17 +29,17 @@ use strata_primitives::params::Params;
 use strata_rocksdb::{
     broadcaster::db::BroadcastDb, init_broadcaster_database, init_core_dbs,
     init_sequencer_database, open_rocksdb_database, sequencer::db::SequencerDB, CommonDb,
-    DbOpsConfig, RBSeqBlobDb,
+    DbOpsConfig, RBSeqBlobDb, ROCKSDB_NAME,
 };
 use strata_rpc_api::{StrataAdminApiServer, StrataApiServer, StrataSequencerApiServer};
 use strata_status::StatusChannel;
 use strata_storage::{
-    managers::{create_db_managers, DbManagers},
+    managers::{create_db_manager, DbManager},
     ops::bridge_relay::BridgeMsgOps,
     L2BlockManager,
 };
 use strata_sync::{self, L2SyncContext, RpcSyncPeer};
-use strata_tasks::{init_task_manager, ShutdownSignal, TaskExecutor};
+use strata_tasks::{ShutdownSignal, TaskExecutor, TaskManager};
 use tokio::{
     runtime::Handle,
     sync::{broadcast, oneshot},
@@ -47,7 +49,6 @@ use tracing::*;
 use crate::{args::Args, helpers::*};
 
 mod args;
-mod config;
 mod errors;
 mod extractor;
 mod helpers;
@@ -81,7 +82,12 @@ fn main_inner(args: Args) -> anyhow::Result<()> {
         .map_err(anyhow::Error::from)?;
 
     // Init the task manager and logging before we do anything else.
-    let task_manager = init_task_manager();
+    let runtime = tokio::runtime::Builder::new_multi_thread()
+        .enable_all()
+        .thread_name("strata-rt")
+        .build()
+        .expect("init: build rt");
+    let task_manager = TaskManager::new(runtime.handle().clone());
     let executor = task_manager.executor();
 
     init_logging(executor.handle());
@@ -91,12 +97,12 @@ fn main_inner(args: Args) -> anyhow::Result<()> {
     let pool = threadpool::ThreadPool::with_name("strata-pool".to_owned(), 8);
 
     // Open and initialize rocksdb.
-    let rbdb = open_rocksdb_database(&config.client.datadir)?;
+    let rbdb = open_rocksdb_database(&config.client.datadir, ROCKSDB_NAME)?;
     let ops_config = DbOpsConfig::new(config.client.db_retry_count);
 
     // Initialize core databases
     let database = init_core_dbs(rbdb.clone(), ops_config);
-    let managers = create_db_managers(database.clone(), pool.clone());
+    let manager = create_db_manager(database.clone(), pool.clone());
 
     // Set up bridge messaging stuff.
     // TODO move all of this into relayer task init
@@ -104,7 +110,7 @@ fn main_inner(args: Args) -> anyhow::Result<()> {
     let bridge_msg_ctx = strata_storage::ops::bridge_relay::Context::new(bridge_msg_db);
     let bridge_msg_ops = Arc::new(bridge_msg_ctx.into_ops(pool.clone()));
 
-    let checkpoint_handle: Arc<_> = CheckpointHandle::new(managers.checkpoint()).into();
+    let checkpoint_handle: Arc<_> = CheckpointHandle::new(manager.checkpoint().clone()).into();
     let bitcoin_client = create_bitcoin_rpc_client(&config)?;
 
     // Check if we have to do genesis.
@@ -121,7 +127,7 @@ fn main_inner(args: Args) -> anyhow::Result<()> {
         &config,
         params.clone(),
         database,
-        &managers,
+        &manager,
         bridge_msg_ops,
         bitcoin_client,
     )?;
@@ -156,7 +162,7 @@ fn main_inner(args: Args) -> anyhow::Result<()> {
             let sequencer_rpc = &fullnode_config.sequencer_rpc;
             info!(?sequencer_rpc, "initing fullnode task");
 
-            let rpc_client = executor.handle().block_on(sync_client(sequencer_rpc));
+            let rpc_client = runtime.block_on(sync_client(sequencer_rpc));
             let sync_peer = RpcSyncPeer::new(rpc_client, 10);
             let l2_sync_context = L2SyncContext::new(
                 sync_peer,
@@ -296,7 +302,7 @@ fn start_core_tasks(
     config: &Config,
     params: Arc<Params>,
     database: Arc<CommonDb>,
-    managers: &DbManagers,
+    managers: &DbManager,
     bridge_msg_ops: Arc<BridgeMsgOps>,
     bitcoin_client: Arc<BitcoinClient>,
 ) -> anyhow::Result<CoreContext> {
@@ -307,7 +313,7 @@ fn start_core_tasks(
         config,
         database.clone(),
         params.as_ref(),
-        managers.l2(),
+        managers.l2().clone(),
         executor.handle(),
     )?;
 
@@ -355,7 +361,7 @@ fn start_core_tasks(
         pool,
         params,
         sync_manager,
-        l2_block_manager: managers.l2(),
+        l2_block_manager: managers.l2().clone(),
         status_channel,
         engine,
         relayer_handle,
