+++ crates/btcio/src/reader/config.rs
@@ -1,5 +1,6 @@
 use std::sync::Arc;
 
+use strata_common::config::Config;
 use strata_primitives::params::Params;
 
 #[derive(Clone, Debug)]
@@ -15,6 +16,14 @@ pub struct ReaderConfig {
 }
 
 impl ReaderConfig {
+    pub fn from_config_and_params(config: Config, params: Arc<Params>) -> Self {
+        Self {
+            max_reorg_depth: config.sync.max_reorg_depth,
+            client_poll_dur_ms: config.sync.client_poll_dur_ms,
+            params,
+        }
+    }
+
     pub fn new(max_reorg_depth: u32, client_poll_dur_ms: u32, params: Arc<Params>) -> Self {
         Self {
             max_reorg_depth,
