+++ crates/common/Cargo.toml
@@ -4,9 +4,15 @@ name = "strata-common"
 version = "0.1.0"
 
 [dependencies]
+bitcoin.workspace = true
 opentelemetry.workspace = true
 opentelemetry-otlp.workspace = true
 opentelemetry_sdk.workspace = true
+serde.workspace = true
+strata-primitives.workspace = true
 tracing.workspace = true
 tracing-opentelemetry.workspace = true
 tracing-subscriber.workspace = true
+
+[dev-dependencies]
+toml.workspace = true
