+++ crates/common/src/lib.rs
@@ -1,4 +1,5 @@
 //! Crate includes reusable utils for services that handle common behavior.
 //! Such as initializing the tracing framework and whatever else.
 
+pub mod config;
 pub mod logging;
