+++ crates/consensus-logic/src/sync_manager.rs
@@ -8,7 +8,7 @@ use strata_db::traits::Database;
 use strata_eectl::engine::ExecEngineCtl;
 use strata_primitives::params::Params;
 use strata_status::StatusChannel;
-use strata_storage::managers::DbManagers;
+use strata_storage::managers::DbManager;
 use strata_tasks::TaskExecutor;
 use tokio::sync::{broadcast, mpsc};
 
@@ -79,7 +79,7 @@ pub fn start_sync_tasks<
 >(
     executor: &TaskExecutor,
     database: Arc<D>,
-    managers: &DbManagers,
+    managers: &DbManager,
     engine: Arc<E>,
     pool: threadpool::ThreadPool,
     params: Arc<Params>,
@@ -97,7 +97,7 @@ pub fn start_sync_tasks<
     // Start the fork choice manager thread.  If we haven't done genesis yet
     // this will just wait until the CSM says we have.
     let fcm_database = database.clone();
-    let fcm_l2_block_manager = managers.l2();
+    let fcm_l2_block_manager = managers.l2().clone();
     let fcm_engine = engine.clone();
     let fcm_csm_controller = csm_controller.clone();
     let fcm_params = params.clone();
@@ -122,9 +122,9 @@ pub fn start_sync_tasks<
     let client_worker_state = worker::WorkerState::open(
         params.clone(),
         database,
-        managers.l2(),
+        managers.l2().clone(),
         cupdate_tx,
-        managers.checkpoint(),
+        managers.checkpoint().clone(),
     )?;
 
     let csm_engine = engine.clone();
