+++ crates/storage/src/managers/mod.rs
@@ -6,21 +6,22 @@ pub mod checkpoint;
 pub mod l2;
 
 /// A consolidation of database managers and ops. Ideally ops should not be exposed.
-pub struct DbManagers {
+#[derive(Clone)]
+pub struct DbManager {
     l2_block_manager: Arc<l2::L2BlockManager>,
     checkpoint_manager: Arc<checkpoint::CheckpointDbManager>,
 }
-impl DbManagers {
-    pub fn l2(&self) -> Arc<l2::L2BlockManager> {
-        self.l2_block_manager.clone()
+impl DbManager {
+    pub fn l2(&self) -> &Arc<l2::L2BlockManager> {
+        &self.l2_block_manager
     }
 
-    pub fn checkpoint(&self) -> Arc<checkpoint::CheckpointDbManager> {
-        self.checkpoint_manager.clone()
+    pub fn checkpoint(&self) -> &Arc<checkpoint::CheckpointDbManager> {
+        &self.checkpoint_manager
     }
 }
 
-pub fn create_db_managers<D>(db: Arc<D>, pool: threadpool::ThreadPool) -> DbManagers
+pub fn create_db_manager<D>(db: Arc<D>, pool: threadpool::ThreadPool) -> DbManager
 where
     D: Database + Sync + Send + 'static,
 {
@@ -29,7 +30,7 @@ where
         db.clone(),
     ));
     let l2_block_manager = Arc::new(l2::L2BlockManager::new(pool.clone(), db.clone()));
-    DbManagers {
+    DbManager {
         checkpoint_manager,
         l2_block_manager,
     }
