+++ crates/tasks/src/lib.rs
@@ -2,5 +2,5 @@ mod manager;
 mod pending_tasks;
 mod shutdown;
 
-pub use manager::{init_task_manager, TaskError, TaskExecutor, TaskManager};
+pub use manager::{TaskError, TaskExecutor, TaskManager};
 pub use shutdown::{ShutdownGuard, ShutdownSignal};
